+++ Source/CS2/Constants/SceneObjectAttributeNames.h
@@ -1,0 +1,9 @@
+#pragma once
+
+namespace cs2::scene_object_attribute
+{
+
+constexpr auto kSpawnInvulnerabilityHash = 0x244EC9B0; // murmur2 hash of "SpawnInvulnerability"
+constexpr auto kInvulnerabilityColorHash = 0xB2CAF4DF; // murmur2 hash of "InvulnerabilityColor"
+
+}
